feat(Entity): add canonical name
- Add IRepositoryQueryAsyncAdvice
- Add ICanonicalName and AdviceAddCanonicalName

+++ src/Schemata.Entity.LinqToDB/LinQ2DbRepository.cs
@@ -40,7 +40,8 @@ public class LinQ2DbRepository<TContext, TEntity> : RepositoryBase<TEntity>
     public override async IAsyncEnumerable<TResult> ListAsync<TResult>(
         Func<IQueryable<TEntity>, IQueryable<TResult>>? predicate,
         [EnumeratorCancellation] CancellationToken      ct = default) {
-        var enumerable = BuildQuery(predicate).AsAsyncEnumerable().WithCancellation(ct);
+        var query      = await BuildQueryAsync(predicate, ct);
+        var enumerable = query.AsAsyncEnumerable().WithCancellation(ct);
 
         await foreach (var entity in enumerable) {
             ct.ThrowIfCancellationRequested();
@@ -52,54 +53,62 @@ public class LinQ2DbRepository<TContext, TEntity> : RepositoryBase<TEntity>
         Func<IQueryable<TEntity>, IQueryable<TResult>>? predicate,
         CancellationToken                               ct = default)
         where TResult : default {
-        return await BuildQuery(predicate).FirstOrDefaultAsync(ct);
+        var query = await BuildQueryAsync(predicate, ct);
+        return await query.FirstOrDefaultAsync(ct);
     }
 
     public override async Task<TResult?> SingleOrDefaultAsync<TResult>(
         Func<IQueryable<TEntity>, IQueryable<TResult>>? predicate,
         CancellationToken                               ct = default)
         where TResult : default {
-        return await BuildQuery(predicate).SingleOrDefaultAsync(ct);
+        var query = await BuildQueryAsync(predicate, ct);
+        return await query.SingleOrDefaultAsync(ct);
     }
 
     public override async Task<bool> AnyAsync<TResult>(
         Func<IQueryable<TEntity>, IQueryable<TResult>>? predicate,
         CancellationToken                               ct = default) {
-        return await BuildQuery(predicate).AnyAsync(ct);
+        var query = await BuildQueryAsync(predicate, ct);
+        return await query.AnyAsync(ct);
     }
 
     public override async Task<int> CountAsync<TResult>(
         Func<IQueryable<TEntity>, IQueryable<TResult>>? predicate,
         CancellationToken                               ct = default) {
-        return await BuildQuery(predicate).CountAsync(ct);
+        var query = await BuildQueryAsync(predicate, ct);
+        return await query.CountAsync(ct);
     }
 
     public override async Task<long> LongCountAsync<TResult>(
         Func<IQueryable<TEntity>, IQueryable<TResult>>? predicate,
         CancellationToken                               ct = default) {
-        return await BuildQuery(predicate).LongCountAsync(ct);
+        var query = await BuildQueryAsync(predicate, ct);
+        return await query.LongCountAsync(ct);
     }
 
     public override async Task AddAsync(TEntity entity, CancellationToken ct = default) {
-        await Advices<IRepositoryAddAsyncAdvice<TEntity>>.AdviseAsync(ServiceProvider, entity, ct);
-
         await BeginTransactionAsync(ct);
+
+        var next = await Advices<IRepositoryAddAsyncAdvice<TEntity>>.AdviseAsync(ServiceProvider, entity, ct);
+        if (!next) return;
 
         RowsAffected += await Context.InsertAsync(entity, TableName, token: ct);
     }
 
     public override async Task UpdateAsync(TEntity entity, CancellationToken ct = default) {
-        await Advices<IRepositoryUpdateAsyncAdvice<TEntity>>.AdviseAsync(ServiceProvider, entity, ct);
-
         await BeginTransactionAsync(ct);
+
+        var next = await Advices<IRepositoryUpdateAsyncAdvice<TEntity>>.AdviseAsync(ServiceProvider, entity, ct);
+        if (!next) return;
 
         RowsAffected += await Context.UpdateAsync(entity, TableName, token: ct);
     }
 
     public override async Task RemoveAsync(TEntity entity, CancellationToken ct = default) {
-        await Advices<IRepositoryRemoveAsyncAdvice<TEntity>>.AdviseAsync(ServiceProvider, entity, ct);
-
         await BeginTransactionAsync(ct);
+
+        var next = await Advices<IRepositoryRemoveAsyncAdvice<TEntity>>.AdviseAsync(ServiceProvider, entity, ct);
+        if (!next) return;
 
         RowsAffected += await Context.DeleteAsync(entity, TableName, token: ct);
     }
@@ -123,9 +132,16 @@ public class LinQ2DbRepository<TContext, TEntity> : RepositoryBase<TEntity>
         return rows;
     }
 
-    private IQueryable<TResult> BuildQuery<TResult>(Func<IQueryable<TEntity>, IQueryable<TResult>>? predicate) {
+    private async Task<IQueryable<TResult>> BuildQueryAsync<TResult>(
+        Func<IQueryable<TEntity>, IQueryable<TResult>>? predicate,
+        CancellationToken                               ct) {
         var table = Context.GetTable<TEntity>().TableName(TableName).AsQueryable();
-        return BuildQuery(table, predicate);
+
+        var query = new QueryContainer<TEntity>(table);
+
+        await Advices<IRepositoryQueryAsyncAdvice<TEntity>>.AdviseAsync(ServiceProvider, query, ct);
+
+        return BuildQuery(query.Query, predicate);
     }
 
     private async Task BeginTransactionAsync(CancellationToken ct) {
